renamed old members section to members-v1
